services

+++ src/main/java/com/ITtraining/project/controllers/CategoryController.java
@@ -1,4 +1,4 @@
-package com.example.project.controllers;
+package com.ITtraining.project.controllers;
 
 import java.util.List;
 
@@ -9,8 +9,10 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.example.project.entities.CategoryEntity;
-import com.example.project.repositories.CategoryRepository;
+import com.ITtraining.project.entities.CategoryEntity;
+import com.ITtraining.project.repositories.CategoryRepository;
+import com.ITtraining.project.services.BillDao;
+import com.ITtraining.project.services.OfferDao;
 
 @RestController
 @RequestMapping(value = "/api/v1/project/categories")
@@ -19,13 +21,19 @@ public class CategoryController {
 	@Autowired
 	private CategoryRepository categoryRepo;
 
-	// vraca sve kategorije
+	@Autowired
+	private BillDao billDao;
+
+	@Autowired
+	private OfferDao offerDao;
+
+	// find all categories
 	@RequestMapping
 	public List<CategoryEntity> getAllCategories() {
 		return (List<CategoryEntity>) categoryRepo.findAll();
 	}
 
-	// postavljanje nove kategorije
+	// add new category
 	@RequestMapping(method = RequestMethod.POST)
 	public CategoryEntity addNewCategory(@RequestBody CategoryEntity newCategory) {
 
@@ -36,7 +44,7 @@ public class CategoryController {
 		return categoryRepo.save(newCategory);
 	}
 
-	// izmena postojece kategorije
+	// modify an existing category
 	@RequestMapping(value = "/{id}", method = RequestMethod.PUT)
 	public CategoryEntity updateCategory(@PathVariable Integer id, @RequestBody CategoryEntity category) {
 
@@ -58,22 +66,26 @@ public class CategoryController {
 		return null;
 	}
 
-	// brisanje postojece kategorije
+	// delete an existing category
 	@RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
 	public CategoryEntity deleteCategory(@PathVariable Integer id) {
 
 		if (categoryRepo.existsById(id)) {
 
 			CategoryEntity categoryEntity = categoryRepo.findById(id).get();
-			categoryRepo.deleteById(id);
 
-			return categoryEntity;
+			if (billDao.findActiveBillsForCategory(categoryEntity).size() == 0
+					&& offerDao.findActiveOffersForCategory(categoryEntity).size() == 0) {
+
+				categoryRepo.deleteById(id);
+				return categoryEntity;
+			}
 		}
 
 		return null;
 	}
 
-	// vraca kategoriju po Id-u
+	// find category by Id
 	@RequestMapping(value = "/{id}")
 	public CategoryEntity getCategoryById(@PathVariable Integer id) {
 		return categoryRepo.findById(id).get();
